Fix use of orig_nents rather than nents.
The sg_table struct has two fields that record the number of entries
(nents and orig_nents), and we were using the wrong one in these
cases.

The DMA mapping operation may change the number of entries, which is
reflected in nents but orig_nents is not updated.

We are using nents further down the code to loop over the entries, so we
should only be allocating the space for this number.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>