Fix double-mapping of dmabuf
We were passing the allocator device to dma_buf_attach, which leads to
the linux dma framework attempting to map the buffer using the iommu
mentioned in the dts for the allocator, which is not what we want
because we are handling the mapping ourselves. This was leading to two
mappings occurring which was leading to crashes and corrupted data.

This patch passes the ethosn_core device instead, which does not have an
associated iommu, so the linux dma framework does a "direct" mapping,
which doesn't seem to cause any problems.

It remains unclear if we are handling buffer mapping properly, as it
seems this should be done using the linux dma framework. However that
would be a larger change and is out of scope for this small fix.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>