Fix incorrect freeing of pages
In the case of dmabuf importing, we do not allocate pages in the iommu.
Instead we copy the page structs from the sg_table returned by the
dmabuf attaching. However the code was still freeing these pages when
the buffer is destroyed, which is incorrect.

This patch makes it so that we only free the pages if we allocated
them.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>