Remove non-process memory allocator APIs
All components above the kernel module in the driver stack have been
switched over to use the new process memory allocator when creating
networks and buffers. As process memory allocators are required to
implement memory isolation, the old methods which don't use the
process memory allocator for creation of networks and buffers have
been removed in this patch.

The major version of kernel module has been bumped with this change as
it is not backwards compatible.

Signed-off-by: Rajasekaran Kalidoss <rajasekaran.kalidoss@arm.com>