Fix mismatching checks for plans fitting in SRAM.
Some plans do not contain ple operations (e.g. DMA only operations like
Reshape). And thus do not need to account for a PLE kernel in SRAM.

The calculation which finds the best plans did not account for PLE
kernels but when the plan are added to the set of plans it adds a PLE
kernel. This means we could choose a plan which just barely fits
in SRAM but won't be able to add it to the set of plans.

Fix the calculation when adding to the set of plans to check if a PLE
kernel is necessary.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>