Fix firmware not released on failed firmware load
When the kernel module's firmware_load function fails to find a
compatible firmware in the big firmware binary or the firmware has an
incompatible version, it returns directly without releasing the firmware
instance. This causes the firmware instance to stay in the kernel's
cache and later requests for the firmware will return the same cached
instance regardless if the firmware binary has changed in the
filesystem.

The error handling has now been fixed so that the firmware instance is
always released.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>