Add magic number to NPU firmware binary
To be able to distinguish the NPU firmware binary from other binaries, a
magic number has been added to the start of the binary that the kernel
module will use to identify the binary before attempting to retrieve
other information from it.

The change to the firmware binary's format is not backwards compatible
so the firmware's major version has been bumped with this change.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>