Fix warnings about function name usage in logs
checkpatch is generating warnings for the iommu_iova_map and
iommu_import functions in the kernel module because they write the
function name directly in log strings instead of using a "%s" format and
passing __func__ which is the preferred way.

The logs have been updated accordingly to fix the warnings.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>