Fix IFM stripe delta calculations
Before this patch, the command stream contained two IFM delta fields -
"default" and "edge". This is not enough to encode the different values
 of IFM delta needed across the whole tensor for all supported cases.

In particular, when there is only a small amount of IFM data in the
final stripe, the delta should be set to a smaller value than
ifmDeltaDefault for the penultimate OFM stripe. This was not accounted
for.

A new field "ifmDeltaOneFromEdge" has been added to support these cases,
and the code in the support library which calculates all the delta
values has been updated and simplified.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>