Set ethosn_global_device_for_testing to first dev
In a multi-device system, the driver's probe function is called more
than once and this meant that ethosn_global_device_for_testing was
being set to the last device. However the test code was using the first
device, and this mismatch led to many failing tests.

This patch changes the behaviour of the kernel module so that
ethosn_global_device_for_testing is set to the first device, rather
than the last so that it is consistent with the test code.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>