Change kernel module UAPI to use __kernel_size_t
The Linux Kernel's userspace headers does not provide the size_t type so
in order to use the kernel module's UAPI, the user must include
"stddef.h" before the UAPI header or size_t will be an unknown type. To
ensure that a type of the same size is used and that no extra headers
are needed to use the UAPI. The use of size_t in the UAPI header has
been changed with __kernel_size_t.

Inside the kernel, size_t is an alias for __kernel_size_t so the kernel
module code doesn't need to be changed.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>