Add protected flag to proc_mem_allocator ioctl
As preparation for kernel module support to run protected inferences
the ETHOSN_IOCTL_CREATE_PROC_MEM_ALLOCATOR ioctl has been updated to
take a parameter struct that indicates if it will be used for protected
or non-protected inferences. All existing use of the ioctl must be
updated to include this.

The actual usage of the parameter will be added in later patch.

Signed-off-by: Bjorn Engstrom <bjoern.engstroem@arm.com>