Add usage of is_protected flag in proc_mem_allocator
The kernel module must be aware when protected inferences are going to
be performed so that it can do the necessary configuration of the NPU
using the Ethos-N SiP service and to only allow imported input,
intermediate and output buffers to be used.
This is be done by adding a "is_protected" flag to the per-process
memory protection context creation. A context created with the
is_protected flag set can only be used for protected inferences and
vice-versa.

Solves: NNXSW-4228
Signed-off-by: Bjorn Engstrom <bjoern.engstroem@arm.com>