Prevent zero byte DMA allocations
Zero-byte allocations are not very meaningful and are likely to cause
subtle bugs. This patch adds validation that prevents this in the
ethosn_dma_alloc() function.

The network constant_dma_data was currently relying on a zero-byte
allocation in cases where the network has no weights, which has been
changed in this patch so that no allocation is made in this case.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>