Add logging and error checking for debugfs
There is an extant bug where insmod'ing the kernel module doesn't
correctly populate the debugfs. This patch aims to fix this by removing
the debugfs nodes when an error occurs loading the kernel module, so
that we don't leave them behind.

It also adds some logging when debugfs_create_dir fails, and causes the
insmod to fail in this case. This should make future issues easier to
debug.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>