Speculative fix and debugging for read_reg crash
Several intermittent crashes have been observed that appear to be due to
ethosn_read_top_reg failing to read the register, possibly due to a
NULL pointer.

This patch changes the device registered with the interrupt handler to
the core, rather than the parent device, as this is the object that is
needed in the interrupt handler, and so a mismatch of lifetimes could
have been causing the crash.

It also adds some WARN_ON macros to better log a possible NULL pointer
in case we see the issue again.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>