Fix allowed pids for process memory allocator
Current implementation has no check on pid range during
ethosn_process_mem_allocator_create, allowing creation of an
ethosn_allocator with zero or negative pid, but (a sub-function of)
proc_mem_allocator_release requires a positive pid in order to fully
release the allocator. This results in kref_put not decrementing kref.

In addition to range check during creation, this patch adds a check to
make sure the value of kref_put is checked okay.

Signed-off-by: Bjorn Engstrom <bjoern.engstroem@arm.com>