Fix nullptr dereference in sub allocator clean up
If the probing fails before the carveout allocator has been created, the
sub allocator clean up code will crash because it will dereference the
null pointer returned when trying to get a sub allocator from a top
allocator that doesn't exist.

To resolve this, the clean up code will now check that the pointer is
not null before dereferencing it.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>