Move stacks and vector table into firmware binary
Previously, the stacks (both privileged and unprivileged) and the vector
table used by the firmware were allocated by the kernel and the
firmware was told where these are at boot time. This patch changes this
so that the stacks and vector table are instead built into the firmware
binary and the kernel does not allocate these explicitly.

This change is needed so that the booting process that the TF-A will do
when loading a protected firmware can be made simpler.

The kernel module no longer creates and fills in the vector table, or
provides a stack pointer to the firmware. When it boots the firmware,
it simply points the INITVTOR to the vector table in the firmware
binary.

The firmware binary is loaded as one big blob into memory, but is mapped
into the SMMU with different protection flags for the PLE code and
vector table (read-only) and the rest(read-write).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>