Fix lifetime and concurrency bugs in kernel module
Lock core mutex in ethosn_pm_common_suspend, as we're querying/modifying
the state of the core.

Lock core mutex in inference_release as we're querying the state of the
inference, which might be changed by the inference finishing in the
background.

Rename ethosn_network_poll to ethosn_set_inference_done.

Check core->current_inference in ethosn_set_inference_done, to account
for cases where the inference finishes after it has already been
cancelled.

Remove ref-counting of ethosn_inference struct - this isn't needed
because the user-space file handles already ref-count the lifetime of
the inference.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>