Add core reset when inference context changes
The nsaid's for the different streams must be set up based on if a
protected or non-protected is to be executed. This is done during core
reset and currently this is only done (before an inference is
scheduled) if the alloc id differs.
This patch will add a flag to the core to keep track of if it was set
up for protected or non-protected context, and force a restart if the
context differs.

Signed-off-by: Bjorn Engstrom <bjoern.engstroem@arm.com>