Simplify profiling time sync
Remove the time sync message that was previously sent between the
firmware and kernel module. The synchronisation is now done as part of
the profiling configuration ack which is already sent. The firmware
sends its timestamp along with the ack, which the kernel compares to
its own and exposes the difference as a debugfs file alongside the
existing firmware_profiling_buffer file.

The driver library reads this file to get the offset.

This is simpler because there is less messages exchanged and no
encoding/decoding of the timestamp into multiple messages

Signed-off-by: Rob Hughes <robert.hughes@arm.com>