Update Arm NN so that deregister allocators is called in teardown
If you run test using dma buff more the once in one go the registered
allocators are still registered the 2nd time the test it setup casuing
the test to fail. This is solved by always calling
DeregisterAllocator() in the teardown to make sure it is removed
before adding on in the text test during it's setup. This was fixed
by Arm NN.

Signed-off-by: Zingo Andersen <zingo.andersen@arm.com>