Speed up kernel module IOMMU domain clean up
When an SMMU is present, the kernel module must unmap all pages from the
IOMMU domains before the kernel module is unloaded or if an issue occurs
while setting up the pages for an IOMMU domain, all mapped pages must be
unmapped. This is currently done by using the iommu_unmap function which
will flush the TLB for each unmap call, which makes the process of
unmapping the pages very slow.

To speed up the process of unmapping the pages in these cases, the
iommu_unmap_fast function is now used instead as it won't flush the TLB.
This is instead the caller's responsibility so the kernel module will
flush the TLB after all the pages have been unmapped.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>