Enable FCAF for more cases
Previously we required that the stripe shape had to be a multiple of the
FCAF cell size in all dimensions, but this is more restrictive than
necessary. It is fine to use FCAF if the stripe shape isn't a multiple
of the cell size in a particular dimension, if there is only a single
stripe in that dimension.

In order to not break the non-cascading compiler, a copy of the old
implementation has been left behind. The non-cascading code would
require other changes to round up the stripe shape to a multiple of the
FCAF cell in order to make this work, and this is out of scope.

Some unit tests have been temporarily removed but will be added back in
the following commit, once some other changes have been applied.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>