Add temporary debugging instrumentation
Store the address about to be read into a register so that it can be
inspected in the crash dump, should the read cause a crash. By default,
the compiler overwrites the address with the read value, which means we
can't inspect the address after the crash happens(the crash seems to
occur immediately _after_ the value is read).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>