Fix shared interrupt handling during NPU reset
When the NPU is on a shared interrupt line, the NPU's interrupt handler
will be called for interrupts coming from other hardware on the same
line and needs to check a status register to see if the interrupt was
from the NPU. If an interrupt occurs while the NPU is being reset and
the NPU is in secure mode, the kernel will panic because the interrupt
handler will try to access the status register before it has been
delegated to the non-secure world by the TF-A SiP service.

To handle this, the kernel module will now keep track of when the NPU is
being reset and ignore interrupts until the NPU has been configured to
be used again.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>