Merge channel selector layers when possible
In some cases the support library needs to insert MceParts into the
GraphOfParts which are layers which simply add or remove channels,
rather than performing any arithmetic. This is mostly for padding
channels for split and concat. These new parts can be executed more
efficiently by merging them with a preceding/following McePart in some
circumstances. This patch performs this merging where it is possible
and leads to less computation.

Add the concept of "channel selectors" - Parts can advertise themselves
as channel selectors, and Parts can optionally support merging
themselves with an adjacent channel selector. This is used by the new
method MergeChannelSelectors() which looks through the graph for
channel selectors and attempts to merge them with adjacent parts.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>