Change NPU firmware to only contain one binary
Support for having multiple binaries packed into the NPU firmware binary
has been removed because it is no longer needed and it makes the
handling of the firmware easier. The kernel module's handling of the
firmware has been updated accordingly.

The change to the NPU firmware binary's format is not backwards
compatible so the firmware's major version has been bumped with this
change.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>