Update SRAM/DRAM DMA check to use DRAM shape too
IsSramBufferCompatibleWithDramFormat is used to check if an SRAM buffer
can be DMA'd to/from a DRAM buffer. Previously it was assuming that the
shape of the SRAM and DRAM buffers were the same, but this is not always
the case and we want to be able to use this method for those cases too.

This patch extends the function to take extra parameters so that it can
perform these checks too.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>