Add kernel TZMP1 restrictions and firmware check
Add build flag (#define ETHOSN_TZMP1) to the kernel module to indicate
that TZMP1 (protected firmware) is used. This is mutually exclusive to
the existing ETHOSN_NS (non-secure) flag.

Creation of per-process memory allocator to run protected inferences is
not allowed unless the kernel module is built for TZMP1.

Carveout is not supported in a kernel module built for TZMP1.

In addition to the existing handshake between the kernel module and SiP
service, the kernel queries the NPU's firmware version from the SiP
service to validate that it is compatible

Signed-off-by: Rob Hughes <robert.hughes@arm.com>