Use numSrams rather than numOgs for depthwise
When doing depthwise convolution, only one OG per CE is used. This means
that the minimum feasible stripe depth is based on the number of SRAMs
rather than the number of OGs.

This change is necessary to support max pooling, as max pooling requires
stripes with a depth of only one channel per PLE, which was previously
not the case when combined with a depthwise.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>