Prevent max pool 3x3 2,2 from having >1 channel per PLE
The max pool 3x3 2,2 PLE kernel only supports stripes with one channel
per PLE, when it is also splitting in height. This limitation was not
being honoured in the support library, so we would generate invalid
command streams in some cases.

This patch adds this limitation to the support library so that it will
not generate such plans.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>