Fix rounding of FCAF stripes
Previously we were rounding edgeStripeShape width and height to the FCAF
cell shape if the format is FCAF. This was causing problems because the
stripe layout then didn't match the layout the PLE was assuming, which
rounds to the brick group shape(8). This is a problem for FCAF_WIDE
where the cell width is 16, and the DMA was then using assuming
different layout to the PLE, leading to incorrect results.

This patch fixes this by only rounding to the brick group shape for
edgeStripeShape width & height, as it is not necessary to round all the
way up to the FCAF cell size because partial width/height work fine.

This also required fixing the CalculateTileSize to make sure there is
space for FCAF to be decompressed into the last slot in the tile,
because the HW always writes to SRAM in full FCAF cell size if the
source is FCAF compressed, and this was leading to issues where the DMA
would overwrite beyond the end of the last slot in the tile.

This caused some unit tests to fail because slightly less plans are now
valid. These tests are no longer relevant and so have been removed.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>