Change address extension setup based on NPU mode
Currently the kernel module will setup the address extension for stream0
and ask the firmware to setup the address extension for stream1 and 2.
This has now been changed so the SiP service will setup the address
extensions when the driver stack is built for a NPU in secure or TZMP1
mode. When the driver stack is built for a non-secure NPU, the SiP
service is not available so the kernel module will setup all the address
extensions.

The support to setup stream1 and 2 has been removed from the firmware
and the firmware version has been bumped to indicate this change.

When the driver stack is built for a NPU in secure or TZMP1 mode, a new
SiP service version that supports setting up the address extensions
registers is required.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>