Attempt to fix perf regression from FCAF rounding
The previous patch "Fix rounding of FCAF stripes" introduced some
rounding up of tile sizes to account for FCAF_WIDE. Whilst correct,
this resulted in more SRAM being used and this led to performance
regressions on some networks.

This patch changes the rounding so that it is only performed if it won't
use up a significant amount of SRAM. If it would, then instead of
rounding it we mark that buffer as not compatible with FCAF_WIDE.
Although FCAF_WIDE then can't be used with that buffer, FCAF_DEEP
hopefully still can.

This should result in an always-valid graph (either the tile is rounded
up so that FCAF_WIDE can be used, or marked such that FCAF_WIDE won't
be used), that doesn't noticeably hurt performance.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>