Change event and aux setup based on NPU mode
Currently the setup of the NPU's event and aux control registers are
handled by the kernel and firmware. This has now been changed so the SiP
service will set them up when the driver stack is built for a NPU in
secure or TZMP1 mode. When the driver stack is built for a non-secure
NPU, the SiP service is not available so the kernel module will handle
the setup.

The firmware will no longer support setting up the NPU's event and aux
control registers and the firmware version has been bumped to indicate
this change.

When the driver stack is built for a NPU in secure or TZMP1 mode, a new
SiP service version that supports setting up the NPU's event and aux
control registers is required.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>