Set sensible EstimationOptions for compilation
Even when compiling a network (rather than estimating), the cascading
support library still needs an EstimationOptions struct. Previously we
were using a default-constructed struct, which sets the activation
compression savings to 0.0. This meant that we weren't preferring FCAF
as much as we should have been.

This fix set the value of activation compression savings to 0.5 for
compilation, which means that we will prefer FCAF more.

This leads to some performance improvements across the supported
networks.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>