Change ProcMemAllocator constructor to take const char *
The ProcMemAllocator takes std::string as a parameter, passing a const
char * gets implicitly converted to bool (as it is a pointer). This
means the bool constructor takes priority over the std::string
conversion.

e.g. ProcMemAllocator alloc("test") called the bool constructor.
Although this is correct and in the C++ spec it is not intuitive.

Change to use const char * instead of std::string to reduce user error.
std::string features aren't needed in this function.

Bump the driver library version as this is a breaking change.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>