Split Buffer type into Dram/SramBuffer types
The Buffer type was previously being used to represent all the different
kinds of buffers, which led to many fields only being relevant
depending on the value on which kind of buffer it was, leading to more
confusing, more complicated and more error-prone code.

This patch splits the type into different classes, each of which
inherits from the base Buffer class which still contains common
properties.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>