Fix potential overflow in MergeWithChannelSelector
When calculating the number of merged MACs per pixel, multiple 32-bit
values are multiplied and the result is put in a 64-bit integer. The
32-bit values will be multiplied by using a 32-bit multiplication that
may overflow before the value is put in the 64-bit integer.

To ensure that the full 64-bit size is used for the result, the first
value is now cast to 64-bit so a 64-bit multiplication is performed
instead.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>