Remove unnecessary streams in IOMMU sub allocator
Each IOMMU sub allocator will only handle a single stream so having a
stream instance for each stream type in the internal struct and using a
lookup function to get the correct stream instance in the struct is
unnecessary.

The internal struct will now only hold a single stream instance and a
field to indicate the stream type. This simplifies the IOMMU sub
allocator code and the allocator API.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>