Get protected NPU firmware info during probing
In a TZMP1 setup, the kernel module will not load the firmware that the
NPU should use. Instead, the NPU will use the firmware that TF-A has
loaded into protected memory.

As a step towards supporting this, when the kernel module is built for
TZMP1, it will now validate that the NPU is setup correctly for TZMP1
and in addition to the NPU firmware version compatibility check that was
previously added, query the TF-A SiP service to get information about
the firmware that will later be used to map it into the SMMU.

The validation of the TZMP1 setup and TF-A SiP service have been moved
to earlier in the probing process to be able to catch potential issues
earlier and to avoid having to repeat checks.

With these changes, a non-secure NPU will no longer be allowed to be
used with a kernel module built for secure or TZMP1 mode because the NPU
setup should always match the kernel module build.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>