Add support to run NPU with protected firmware
In order for the NPU to use the firmware loaded in the protected memory
by TF-A, in a TZMP1 setup, the firmware must be mapped into the SMMU by
the kernel module.

To support this, the IOMMU sub allocator probing has been updated to use
the virtual base addresses provided by the TF-A SiP service and the sub
allocator has been updated to be able to support handling protected
memory and map it into the SMMU.

In a TZMP1 setup, the kernel module will now only map the firmware in
the protected memory into the SMMU and after completing other necessary
setup ask the TF-A SiP service to boot the NPU because non-secure world
access to the register to boot the NPU is no longer delegated by the
TF-A SiP service in a TZMP1 setup.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>