Fix boundary calculation for strided
The boundary field in the dependencies between MCE and PLE was only
checking the first subfilter size, and not the other three. In cases of
striding with padding on the top and left sides, the first filter is in
fact smaller than the other three and so we were miscalculating the
boundary size.

This patch changes this code to use an existing function which uses the
maximum of all the filter sizes.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>