Add firmware load for carveout in core device init
The firmware load in the core device initalization was removed to avoid
loading the firmware twice during the probing. However, in a setup using
carveout memory, the firmware must be loaded in the core device
initalization before anything else has been allocated to ensure that it
is placed first in the reserved carveout memory.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>