Add protected speculative page for TZMP1
When the NPU runs with a firmware from the protected memory, in a TZMP1
setup, the page used to handle speculative reads from the NPU must come
from the same memory because the NPU cannot access any other memory when
using the stream for the firmware.

To support this, the IOMMU sub allocator can now be given a page to use
for speculative accesses when it is created and in the TZMP1 setup, a
page from the protected memory is provided.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>