Set isWideFilter in cascading support library
Some MCE operations need to use the wide filter feature, but we were
never enabling this in the cascading support library. This patch sets
this flag as appropriate. The logic to determine if wide filter is
needed already exists in the weight encoder, and is now exposed from
this in the EncodedWeights struct, and passed through into the command
stream.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>