Fix depthwise depth splitting at end of cascade
Previously we were trying to split the output depth but not the input
depth, which isn't appropriate for depthwise convolution. This patch
changes it so that it won't split either inputs or outputs. We could
instead have split both inputs and outputs, but that would require more
changes as the previous layer of the cascade is not aware of this, so
this approach is simpler for now.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>