Fix Arm NN deprecation warning
The Arm NN warning "The backend makes use of a deprecated interface to
read constant tensors" was being raised each time a convolution,
depthwise or fully connected layer is used with our backend.

This is because we didn't enable the "NonConstWeights" capability in the
backend. This is a misnomer as this capability refers to the ability of
a backend to use weights as inputs, not whether or not the tensor is
constant or dynamic.

We do support weights as inputs but not dynamic tensors. We already
verify the "constness" of weights and biases in the IsSupported checks.

Transpose Convolution still uses the old API as Arm NN haven't updated
it yet.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>