Fix SRAM buffer with multiple consumers
Extend number of dependencies, needed for two consumers of an SRAM
buffer. This doesn't fix the general case, just the case with two
consumers.

Change the glue generation so that DMAs from SRAM are done immediately
in the ending glue rather than waiting until the starting glue of the
next part.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>