Fix packed boundary data with extra IFM stripe
For some valid padding cases when using packed boundary data, the IfmS
will not need to load the final stripe of data on the right/bottom edge
and so the numStripes will be one smaller, but this extra data will
still need to be included the packed boundary data for the
second-to-last row/column.

Add a new field to the command stream for the IfmS to be aware of this,
and tweak the IfmS properties in the command stream generation for this
case.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>