Add FrozenGraphOfParts class
This is an immutable equivalent of GraphOfParts, with faster accessors.
It stores cached versions of all the accessor methods, so it makes the
accessors much faster at the expense of not being able to change any
parts or connections.

This speeds up network compilation, which uses these accessors a lot.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>