Change GetPartOutputs to return unique OutputSlots
Previously, GraphOfParts::GetPartOutputs would return multiple copies of
the same PartOutputSlot if that output slot was connected to multiple
input slots. This was confusing and most places where we called this
had to remove duplicates separately. This patch changes the function to
not return duplicates at all.

This simplifies the code and makes it more amenable to some performance
optimisations in a following patch.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>