Use iteration rather than recursion in Combiner
This results in a roughly 2x speedup of network compilation! It also
reduces the risk of stack overflows and makes it easier to debug and do
performance analysis.

Remove FindBestCombinationForPart and its caching feature. It is
replaced with the new iterative algorithm in Combiner::Run and
Combiner::CalculateSectionsOfAllLengths. See these functions for
details of the new algorithm.

Update Combination class to store a contiguous vector of Elems rather
than a sparse map. This should be faster and fits the usage better.
Store the estimated performance metric here so it doesn't need to be
passed around separately. Remove m_HeadOrderRank and m_PartIdsInOrder
as these are no longer needed.

Add information to SectionContext that used to be passed down through
the recursive StartSection/ContinueSection/EndSection functions.

Make most of Combiner class private/protected to minimise its public
API. Tests that relied on its implementation details are now using a
derived class to re-expose the protected methods. Remove
m_PartOrderTable as this is no longer needed, because the GraphOfParts
is already sorted and contiguously numbered.

Remove StatsType and UpdateStats, as these are no longer relevant.

Change GluePartToCombinationSrcToDests to automatically find all
connected parts, rather than relying on the caller to provide a correct
list of connections.

Some Combiner tests have been removed as they are no longer relevant to
the new design. Others have been updated to use the new interfaces.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>