Encode registers in support library
Previously, the command stream contained metadata (e.g. stripe shapes
etc.) for all agents and the firmware was responsible for using this to
program the registers on the hardware. This logic has been moved to the
support library to speed up the firmware. The command stream now
contains lists of register values rather than agent metadata.

Add DmaRegisters.hpp/cpp, MceRegisters.hpp/cpp and PleRegisters.hpp/cpp
files to the support library which roughly does the same as the old
register encoding in the firmware. They take as input an agent "desc"
which is roughly the same as the old metadata in the command stream,
and outputs a set of register values that should be written by the
firmware. These are then stored into the command stream.

Add RegistersLayout.hpp which defines the encoding of the hardware
registers that we need to provide to the firmware.

Remove agent metadata from command stream. This is now converted to
register values in the support library and so the firmware does not
need this information. Some agent metadata is kept (e.g. PLE kernel
IDs) as they are constant for all stripes of the agent.

The Commands in the command stream now have "extra data" associated with
them, which are stored as four separate arrays (Dma, ProgramMce,
StartMce, Ple) that come after the array of Commands. Each Command has
an optional offset to its extra data, and the type of extra data
depends on the type of command. This design was chosen so that the
commands within a command array remain a fixed size, rather than having
different command types with different types. Each extra data struct is
mostly a set of register values.

The scheduler in the support library can now schedule multiple commands
for the same DMA stripe, if the stripe requires splitting into chunks.
This is because each chunk will need different register values, so need
to be separate commands.

Add tests for register encoding logic in the support library, mostly
moved from the firmware.

Update binary-to-xml conversion to support the updated command stream
format.

Add "AddCascade" helper function to create a CASCADE command. This
handles all the offset calculations to avoid duplicating this in
several different places.

Rename some types which were moved from the command stream to support
library, to prevent name clashes.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>