Fix zero point encoding in dump filename
The zero point can be negative, but it was being implicitly cast to an
unsigned value which resulted in filenames with zero points
like "4294967189" rather than "-12".

Adding new overloads of ToString for signed numbers fixes this.

This led to type conversion warnings that have been fixed by adding
additional overloads for other datatypes too. This includes an overload
for bool, which makes bools appear as True/False in dot file dumps,
which is nicer than 0/1.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>