Support constant layers in cascading compiler
Constant buffers being used as regular inputs to other operations was
previously leading to an assert, as this wasn't expected by the command
stream generator. This patch handles these kind of buffers by adding a
corresponding constant buffer to the BufferManager.

The constant data is also passed through the ConstantPart and DramBuffer
classes so that it can makes its way into the BufferManager.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>